+++ packages/lesswrong/lib/components.ts
@@ -48,7 +48,7 @@ importComponent("NotificationsList", () => require('../components/notifications/
 importComponent("TagRelNotificationItem", () => require('../components/notifications/TagRelNotificationItem'));
 importComponent("NotificationsItem", () => require('../components/notifications/NotificationsItem'));
 importComponent("NotificationsMenuButton", () => require('../components/notifications/NotificationsMenuButton'));
-importComponent("SubscribeTo", () => require('../components/notifications/SubscribeTo'));
+importComponent("NotifyMeButton", () => require('../components/notifications/NotifyMeButton'));
 importComponent("NotificationTypeSettings", () => require('../components/notifications/NotificationTypeSettings'));
 importComponent("NotificationEmailPreviewPage", () => require('../components/notifications/NotificationEmailPreviewPage'));
 importComponent("EmailPreview", () => require('../components/notifications/EmailPreview'));
